Moving registration to own route file
