+++ backend/server.js
@@ -7,7 +7,8 @@ const { log, ExpressAPILogMiddleware } = require('@rama41222/node-logger');
 const mysqlConnect = require('./db');
 //Add Routes
 const routes = require('./routes/routes');
-const users = require('./routes/users')
+const users = require('./routes/users');
+const registration = require('./routes/registration');
 
 // set up some configs for express.
 const config = {
@@ -33,6 +34,7 @@ app.use(ExpressAPILogMiddleware(logger, { request: true }));
 //Add new route locs here
 routes(app, logger);
 users(app, logger);
+registration(app, logger);
 
 // connecting the express object to listen on a particular port as defined in the config object.
 app.listen(config.port, config.host, (e) => {
